requirements.txt
Django>=4.2,<5.0
python-telegram-bot==20.7
asgiref>=3.6.0
sqlparse>=0.4.4
tzdata
